requirements.txt
gitpython
tinydb
python-dotenv
cryptography
tqdm
